Made changes from Code Review
