+++ lib/strategies/api-strategy.js
@@ -12,9 +12,10 @@
  */
 
 const log4js = require("log4js");
-const ServiceConfig = require("./api-strategy-config");
-const TokenUtil = require("./../utils/token-util");
-const PublicKeyUtil = require("./../utils/public-key-util");
+const constants = require('../utils/constants');
+const TokenUtil = require("../utils/token-util");
+const ServiceUtil = require('../utils/service-util');
+const PublicKeyUtil = require("../utils/public-key-util");
 
 const ERROR = {
 	INVALID_REQUEST: "invalid_request", // HTTP 400
@@ -32,7 +33,11 @@ function ApiStrategy(options) {
 	logger.debug("Initializing");
 	options = options || {};
 	this.name = ApiStrategy.STRATEGY_NAME;
-	this.serviceConfig = new ServiceConfig(options);
+	const { TENANT_ID, OAUTH_SERVER_URL } = constants;
+	this.serviceConfig = new ServiceUtil.loadConfig('APIStrategy', [
+		TENANT_ID,
+		OAUTH_SERVER_URL
+	], options);
 	PublicKeyUtil.setPublicKeysEndpoint(this.serviceConfig.getOAuthServerUrl());
 }
