+++ lib/strategies/webapp-strategy.js
@@ -14,11 +14,12 @@
 const log4js = require("log4js");
 const request = require("request");
 const Q = require("q");
-const TokenUtil = require("./../utils/token-util");
-const ServiceConfig = require("./webapp-strategy-config");
-const PublicKeyUtil = require("./../utils/public-key-util");
+const constants = require('../utils/constants');
+const TokenUtil = require("../utils/token-util");
+const ServiceUtil = require('../utils/service-util');
+const PublicKeyUtil = require("../utils/public-key-util");
 const acceptLanguage = require('accept-language');
-const appIdSupportedLanguagesArray =  require("../data/languagesCodesArray");
+const appIdSupportedLanguagesArray = require("../data/languagesCodesArray");
 acceptLanguage.languages(appIdSupportedLanguagesArray);
 
 const STRATEGY_NAME = "appid-webapp-strategy";
@@ -28,7 +29,14 @@ function WebAppStrategy(options) {
 	logger.debug("Initializing");
 	options = options || {};
 	this.name = WebAppStrategy.STRATEGY_NAME;
-	this.serviceConfig = new ServiceConfig(options);
+	const { CLIENT_ID, TENANT_ID, SECRET, OAUTH_SERVER_URL, REDIRECT_URI } = constants;
+	this.serviceConfig = new ServiceUtil.loadConfig('WebAppStrategy', [
+		CLIENT_ID,
+		TENANT_ID,
+		SECRET,
+		OAUTH_SERVER_URL,
+		REDIRECT_URI
+	], options);
 	PublicKeyUtil.setPublicKeysEndpoint(this.serviceConfig.getOAuthServerUrl());
 }
 
@@ -158,7 +166,7 @@ function handleForgotPassword(req, options, strategy) {
 	if (options.successRedirect) {
 		req.session[WebAppStrategy.ORIGINAL_URL] = options.successRedirect;
 	}
-	var  forgotPasswordUrl = generateForgotPasswordUrl(req, strategy);
+	var forgotPasswordUrl = generateForgotPasswordUrl(req, strategy);
 	logger.debug("handleForgotPassword :: redirecting to", forgotPasswordUrl);
 	strategy.redirect(forgotPasswordUrl);
 }
@@ -186,7 +194,7 @@ function handleChangeDetails(req, options, strategy) {
 		if (!error) {
 			if (response.statusCode === 200) {
 				var code = body;
-				var  changeDetailsUrl = generateChangeDetailsUrl(code, req, strategy);
+				var changeDetailsUrl = generateChangeDetailsUrl(code, req, strategy);
 				logger.debug("handleChangeDetails :: redirecting to", changeDetailsUrl);
 				strategy.redirect(changeDetailsUrl);
 			} else {
@@ -214,7 +222,7 @@ function handleRopFlow(options, req, strategy) {
 	const appIdAuthContext = req.session[WebAppStrategy.AUTH_CONTEXT];
 	if (appIdAuthContext && appIdAuthContext.accessTokenPayload["amr"][0] === "appid_anon") {
 		logger.debug("handleRopFlow :: added anonymous access_token to POST body");
-		formData["appid_access_token"] =  appIdAuthContext.accessToken;
+		formData["appid_access_token"] = appIdAuthContext.accessToken;
 	}
 	retrieveTokens(formData, strategy).then(function (appIdAuthContext) {
 		// Save authorization context to HTTP session
@@ -275,7 +283,7 @@ function redirectForAuthorization(req, options, strategy) {
 		return;
 	}
 
-	var stateParameter = { anonymousLogin: false, state: TokenUtil.getRandomNumber()};
+	var stateParameter = { anonymousLogin: false, state: TokenUtil.getRandomNumber() };
 	if (options.allowAnonymousLogin === true) {
 		stateParameter.anonymousLogin = true;
 	}
@@ -397,14 +405,14 @@ function retrieveTokens(formData, strategy) {
 				body = body && JSON.parse(body);
 				let loginErrMsg = "Failed to obtain tokens";
 
-				if (body && (body.error === INVALID_GRANT || body["error_code"] === FORBIDDEN))  {
-					loginErrMsg =  body.error_description;
+				if (body && (body.error === INVALID_GRANT || body["error_code"] === FORBIDDEN)) {
+					loginErrMsg = body.error_description;
 				}
-				let loginError =  new Error(loginErrMsg);
+				let loginError = new Error(loginErrMsg);
 				loginError.statusCode = response && response.statusCode;
 				loginError.code = (body.error || body["error_code"]).toLowerCase();
 				deferred.reject(loginError);
-			}catch (e) {
+			} catch (e) {
 				deferred.reject(new Error("Failed to obtain tokens"));
 			}
 		} else {
@@ -419,13 +427,13 @@ function retrieveTokens(formData, strategy) {
 			};
 
 			Promise.all([TokenUtil.decodeAndValidate(accessTokenString, serviceConfig),
-				TokenUtil.decodeAndValidate(identityTokenString,serviceConfig)]).then(function(tokens){
+				TokenUtil.decodeAndValidate(identityTokenString, serviceConfig)]).then(function (tokens) {
 				if (tokens.length < 2 || !tokens[0] || !tokens[1]) {
 					throw new Error("Invalid access/id token");
 				}
 
 				if (!TokenUtil.validateIssAndAud(tokens[0], serviceConfig)) {
-                    throw new Error("Access token validation failed");
+					throw new Error("Access token validation failed");
 				}
 
 				if (!TokenUtil.validateIssAndAud(tokens[1], serviceConfig)) {
