+++ lib/token-manager/token-manager.js
@@ -15,8 +15,9 @@ const log4js = require("log4js");
 const logger = log4js.getLogger("appid-token-manager");
 const request = require("request");
 
-const ServiceConfig = require("./token-manager-config");
+const constants = require('../utils/constants');
 const TokenUtil = require("../utils/token-util");
+const ServiceUtil = require('../utils/service-util');
 const PublicKeyUtil = require("../utils/public-key-util");
 
 const TOKEN_PATH = '/token';
@@ -24,7 +25,13 @@ const CUSTOM_IDENTITY_GRANT_TYPE = 'urn:ietf:params:oauth:grant-type:jwt-bearer'
 
 function TokenManager(options) {
 	logger.debug("Initializing token manager");
-	this.serviceConfig = new ServiceConfig(options);
+	const { CLIENT_ID, TENANT_ID, SECRET, OAUTH_SERVER_URL } = constants;
+	this.serviceConfig = new ServiceUtil.loadConfig('TokenManager', [
+		CLIENT_ID,
+		TENANT_ID,
+		SECRET,
+		OAUTH_SERVER_URL
+	], options);
 	PublicKeyUtil.setPublicKeysEndpoint(this.serviceConfig.getOAuthServerUrl());
 }
 
@@ -99,6 +106,7 @@ function getTokens(tokenEndpoint, clientId, secret, grant_type, scope, assertion
 }
 
 function validateToken(tokenType, token, serviceConfig) {
+	logger.debug(`Validating ${tokenType} token`);
 	return TokenUtil.decodeAndValidate(token, serviceConfig).then((validatedToken) => {
 		if (!validatedToken) {
 			logger.error(`Invalid ${tokenType} token`);
