+++ lib/utils/constants.js
@@ -1,0 +1,13 @@
+module.exports = {
+	VCAP_SERVICES: "VCAP_SERVICES",
+	VCAP_SERVICES_CREDENTIALS: "credentials",
+	VCAP_SERVICES_SERVICE_NAME1: "AdvancedMobileAccess",
+	VCAP_SERVICES_SERVICE_NAME2: "AppID",
+	VCAP_APPLICATION: "VCAP_APPLICATION",
+	TENANT_ID: "tenantId",
+	CLIENT_ID: "clientId",
+	SECRET: "secret",
+	OAUTH_SERVER_URL: "oauthServerUrl",
+	REDIRECT_URI: "redirectUri",
+	PREFERRED_LOCALE: "preferredLocale"
+};
