+++ lib/utils/service-util.js
@@ -13,29 +13,24 @@
 
 const log4js = require("log4js");
 let logger;
+const {
+	VCAP_SERVICES,
+	VCAP_SERVICES_CREDENTIALS,
+	VCAP_SERVICES_SERVICE_NAME1,
+	VCAP_SERVICES_SERVICE_NAME2,
+	VCAP_APPLICATION,
+	CLIENT_ID,
+	TENANT_ID,
+	SECRET,
+	OAUTH_SERVER_URL,
+	REDIRECT_URI,
+	PREFERRED_LOCALE
+} = require('./constants');
 
-const VCAP_SERVICES = "VCAP_SERVICES";
-const VCAP_SERVICES_CREDENTIALS = "credentials";
-const VCAP_SERVICES_SERVICE_NAME1 = "AdvancedMobileAccess";
-const VCAP_SERVICES_SERVICE_NAME2 = "AppID";
-const VCAP_APPLICATION = "VCAP_APPLICATION";
-const TENANT_ID = "tenantId";
-const CLIENT_ID = "clientId";
-const SECRET = "secret";
-const OAUTH_SERVER_URL = "oauthServerUrl";
-const REDIRECT_URI = "redirectUri";
-const PREFERRED_LOCALE = "preferredLocale";
-
-const requiredParams = {
-	WebAppStrategy: [CLIENT_ID, TENANT_ID, SECRET, OAUTH_SERVER_URL, REDIRECT_URI],
-	APIStrategy: [TENANT_ID, OAUTH_SERVER_URL],
-	TokenManager: [CLIENT_ID, TENANT_ID, SECRET, OAUTH_SERVER_URL]
-};
-
-const checkParams = (configName, config) => requiredParams[configName]
+const checkParams = (configName, requiredParams, config) => requiredParams
 	.map((param) => {
 		if (!config[param]) {
-			logger.error(`Missing ${param} parameter`);
+			throw Error(`Failed to initialize ${configName}. Missing ${param} parameter.`);
 		} else if (param === SECRET) {
 			logger.info(param, '[CANNOT LOG SECRET]');
 		} else {
@@ -45,7 +40,7 @@ const checkParams = (configName, config) => requiredParams[configName]
 	})
 	.every((x) => x);
 
-function loadConfig(configName, options) {
+function loadConfig(configName, requiredParams, options) {
 	logger = log4js.getLogger(`appid-${configName}-config`);
 	logger.debug(`Initializing ${configName} config`);
 	options = options || {};
@@ -53,7 +48,8 @@ function loadConfig(configName, options) {
 	let vcapServiceCredentials = {};
 
 	for (let propName in vcapServices) {
-		if (propName === VCAP_SERVICES_SERVICE_NAME1 || propName === VCAP_SERVICES_SERVICE_NAME2) {
+		// Checks if string starts with the service name
+		if (propName.indexOf(VCAP_SERVICES_SERVICE_NAME1) === 0|| propName.indexOf(VCAP_SERVICES_SERVICE_NAME2) === 0) {
 			vcapServiceCredentials = vcapServices[propName][0][VCAP_SERVICES_CREDENTIALS];
 			break;
 		}
@@ -75,11 +71,12 @@ function loadConfig(configName, options) {
 		}
 	}
 
-	if (!checkParams(configName, serviceConfig)) {
+	if (!checkParams(configName, requiredParams, serviceConfig)) {
 		logger.error(`Failed to initialize ${configName}. Check error logs for more details. All requests to protected endpoints will be rejected`);
 		logger.error(`Ensure your node.js app is either bound to an App ID service instance or pass required parameters in the ${configName} constructor`);
-		throw Error(`Failed to initialize ${configName}. Ensure proper credentials are provided.`);
-	} else if(serviceConfig[PREFERRED_LOCALE]) {
+	}
+
+	if(serviceConfig[PREFERRED_LOCALE]) {
 		logger.info(PREFERRED_LOCALE, serviceConfig[PREFERRED_LOCALE]);
 	}
