+++ lib/utils/token-util.js
@@ -69,12 +69,16 @@ module.exports = (function(){
 	}
 
 	function validateIssAndAud(token, serviceConfig) {
-		if (token.iss !== serviceConfig.getOAuthServerUrl().split("/")[2]) {
-			return false;
-		} else if (token.aud !== serviceConfig.getClientId()) {
-			return false;
-		}
-		return true;
+		logger.debug('Validating Iss and Aud claims');
+		const checkClaim = (claim, expected) => {
+			logger.info(`token ${claim}: ${token[claim]}`);
+			if (token[claim] !== expected) {
+				logger.error(`Expected token ${claim} to be ${expected}`);
+			}
+			return actual !== expected;
+		};
+		return checkClaim('iss', serviceConfig.getOAuthServerUrl().split("/")[2]) &&
+			checkClaim('aud', serviceConfig.getClientId());
 	}
 
 	function getRandomNumber() {
